renomeacao do trafo_potencial para nome antigo
